update the entire sidebar when (re)loading a single feed
The items are aligned, when (re)opening a feed the counts can change. These
counts could resize the sidebar panel and increase/decrease it. This can make
everything dirty in some cases.

The screen will now always look correct, but it may flicker for a small moment
and have some redundant writes.