rework signal handling as event input
This prevents race-conditions in calls which are not async-safe.
Like memory-allocation is not safe (on OpenBSD) in a signal handler
and can cause a recursive lock:

sfeed_curses(56808) in calloc(): recursive call
Abort trap (core dumped)

To reproduce:
* Resize a terminal (SIGWINCH), keep resizing it.
* While resizing spam R (reload all).

Also improve some checks of reading the input after readch() in a EOF
condition.