Fix a crash with small window width and garbage data if utf8pad fails
utf8pad: check the return value, do not print if it returns an error (-1).

updategeom(): clamp width and height to sane values.
The height should be >= 1 (to prevent divide-by-zero).
The width can be set to >= 0.

Remove the check in draw() which prevents drawing. If the window is too small
it will now write the data anyway, although the output will not be nicely
formatted.

To reproduce the crash: make the window width exactly the size of the feed pane
width.  Select the items pane (right-arrow) and scroll down.  The logic of
partial row drawing (which is not in draw()) would be executed anyway and use a
negative value for the pane width.