explicitly define VMIN and VTIME values for read()
VMIN represents the minimum number of bytes that should be received when the
read(2) function successfully returns.

VTIME is a timer of 0.1 second granularity that is used to time out bursty and
short term data transmissions.

This fixes a read buffering issue on OpenIndiana/Illumos and on HaikuOS where
the default reading buffer is 4-5 bytes (noticed by testing).