handle SIGTERM with higher priority while multiple signals are not processed
To reproduce (in this order):

* Open content with sfeed_content (p).
* pkill -SIGTERM sfeed_curses
* pkill -SIGHUP sfeed_curses

Previous behaviour:
The last signal was handled (SIGHUP) and SIGTERM was ignored.

New behaviour:

SIGTERM is treated as more important and is not ignored. SIGHUP is not handled
at all and the program is terminated.

Note that SIGINT is not treated with higher priority. This is intentional.

Note also that the commands (different order):
* pkill -SIGHUP sfeed_curses
* pkill -SIGTERM sfeed_curses

Will still ignore SIGHUP, this makes sense to me since the program is
terminated anyway.