add configurable interactive modes for executed programs
Interactive mode means that the executed program is interactive in the same
terminal.  sfeed_curses will store/restore the terminal attributes and clear
the screen. It will spawn the program. It will then wait until the program is
finished and restore the attributes.
This allows to execute a text-mode browser such as lynx in the same terminal,
like the newsreader newsboat allows too.

In non-interactive mode (was and is the default) it will not change terminal
attributes. Stdout and stderr is suppressed. It will spawn the program.
sfeed_curses will not wait until the program is finished.
This is useful to spawn X11 programs (xdg-open, firefox, etc) in the
background/concurrently.

The interactive mode is configurable using the *_INTERACTIVE environment
variable for all types: PLUMBER, PIPER and YANKER.

Prompted by a question/feedback from Aleksei, thanks!

Misc:
- Ignore SIGINT in the parent program while the program runs, so ^C in lynx is
  handled properly and does not close lynx and sfeed_curses.
- Change pid type from int to pid_t.
- Add 2 usage examples to the README.