UI: truncate the text in the feed pane when the terminal window width is small
getsidebarwidth() calculates the optimal aligned size, but if it does not fit
uses the window size (-1 for the scrollbar).  The counts are still right
aligned (if it fits), the feedname is truncated.

Tested with a bigger fontsize on a Motorola Droid4 (ARM) on Maemo Leste
(Devuan) in vertical-mode (keyboard slide-in) with osso-xterm. This (together
with some local navigation-menu changes) makes it work nicely on this mobile
device.