add keybinds <, > and = to resize feed sidebar pane for small terminals
By default it will still use the optimal consistent size of the longest feed
name.
It can now be manually scaled to a fixed size, '<' decreases, '>' increases.
It can be reset to an automatic optimal size with the '=' keybind.

The value can be hardcoded also more easily if prefered.

Requested by adc and surrounder.