sfeed_content: explicitly specify assumed charset is utf-8 for lynx and close(expr)
The output of sfeed(1) is UTF-8. For the HTML content in sfeed_content now uses
the lynx options: -display_charset="utf-8" -assume_charset="utf-8".

Also:
- Set the command to the expression htmlconv for slightly easier modification.
- Explicitly close the pipe by this expression. This makes sure that when the
  exit; call would be removed the script can be used as a batch content converter
  flushing the output (instead of running it in parallel and mixing the
  output).

To reproduce (on Linux glibc):
curl 'https://globalaffairs.ru/feed/' | sfeed | sfeed_curses and convert the
HTML content with lynx using 'p'.

Lynx will incorrectly detect it as iso-8859-1. This option is not able to be
changed in a fixed way in the config, so specify it on the command-line.

Reported by Aleksei, thanks!