improve dirty flag handling when pane is hidden
This always resets the dirty flag even when the pane is currently hidden.

Do not draw when either the width or height is zero anyway. Document the height
should be > 0, otherwise there could be a divide (or modulo) by zero in certain
functions (like drawing, scrolling).

No functional change intended in the current program logic, but this makes
logically more sense.