fix a memory leak when resizing the window
This happened because setupterm() was called on each resize. ncurses internally
then allocates a new structure every time on calling setupterm(). It also
rereads the terminfo file and entries which is wasteful.

Now it is setup once and only rereads the terminal dimensions using the ioctl
directly.  The global curses variables `columns` and `lines` are not used
anymore.

Note: the contents of the rows and columns are checked, because the ioctl can
succeed, but return zero values, for example on serial consoles.