init: check setupterm() return value and error return value
POSIX specifies:
"If setupterm() detects an error and errret is a null pointer, setupterm()
writes a diagnostic message and exits."

Specify errret and cleanup ourselves. minicurses will ignore this pointer, it
will not use terminfo anyway. Initialize errret to 1 (success).

Make sure to define and check against `OK` or `ERR` for portability, which is
defined in POSIX.

Reference:
https://pubs.opengroup.org/onlinepubs/7908799/xcurses/del_curterm.html

Discussed with quinq, thanks!