clamp height value in getsidebarsizedefault()
This fixes an case in the horizontal layout mode when there are no new items
and only new feeds are shown (using the 't' keybind). The height would be shown
as 1, but internally the sidebar height is set to 0 and it would then require 2
'>' keypresses to increase it from 1 to 2 height.

A similar issue was in commit:
7d9d4d493a90bbd56bbe5cf175970a32ecb915bd

The ranges are checked in adjustsidebarsize() already so remove clamping there.