check if capname is NULL for tparm()
POSIX also does not specify NULL is allowed or not, so assume it isn't.
	https://pubs.opengroup.org/onlinepubs/7908799/xcurses/tigetflag.html

This adds a wrapper to prevent NULL assertion in NetBSD curses/terminfo.

netbsd-curses can use a diagnostic assert and aborts the program:
	https://github.com/oasislinux/netbsd-curses/blob/5b78297a1276a7d64c3be78d1e54251a1203a44a/lib/libterminfo/tparm.c
	http://cvsweb.netbsd.org/bsdweb.cgi/~checkout~/src/lib/libterminfo/tparm.c?rev=1.18&content-type=text/plain&only_with_tag=MAIN

The current ncurses snapshot can show a trace call when the parameter is NULL,
see the tparm_setup() function:
	https://github.com/ThomasDickey/ncurses-snapshots/blob/master/ncurses/tinfo/lib_tparm.c

OpenBSD ncurses (older version), tparam_internal function does nothing if the parameter is NULL:
	http://cvsweb.openbsd.org/cgi-bin/cvsweb/~checkout~/src/lib/libcurses/tinfo/lib_tparm.c?rev=1.9&content-type=text/plain

Noticed on Oasis Linux which uses netbsd-curses and aborts when used in the
terminal with the default $TERM=linux .